refactor: `generate_app()` を分離
