+++ build_util/make_docs.py
@@ -1,6 +1,7 @@
 import json
 from pathlib import Path
 
+from voicevox_engine.app.application import generate_app
 from voicevox_engine.dev.core.mock import MockCoreWrapper
 from voicevox_engine.dev.tts_engine.mock import MockTTSEngine
 from voicevox_engine.preset.PresetManager import PresetManager
@@ -33,11 +34,9 @@ def generate_api_docs_html(schema: str) -> str:
 
 
 if __name__ == "__main__":
-    import run
-
     mock_core = MockCoreWrapper()
     # FastAPI の機能を用いて OpenAPI schema を生成する
-    app = run.generate_app(
+    app = generate_app(
         tts_engines={"mock": MockTTSEngine()},
         cores={"mock": CoreAdapter(mock_core)},
         latest_core_version="mock",
